feat: released avatar
