+++ .changeset/sour-months-punch.md
@@ -1,0 +1,5 @@
+---
+"@brifui/avatar": patch
+---
+
+released avatar
