+++ packages/components/avatar/eslint.config.mjs
@@ -1,0 +1,4 @@
+import { config } from "@brifui/eslint-config/react";
+
+/** @type {import("eslint").Linter.Config} */
+export default config;
