+++ packages/components/avatar/src/avatar.spec.tsx
@@ -1,0 +1,6 @@
+import { render, screen } from "@testing-library/react";
+import { userEvent } from "@testing-library/user-event";
+
+import { Avatar } from "./avatar";
+
+describe("Avatar", () => {})
