+++ packages/components/avatar/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./avatar";
