+++ packages/components/avatar/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@brifui/typescript-config/react.json",
+  "compilerOptions": {
+    "types": ["vitest/globals", "@testing-library/jest-dom"]
+  },
+  "include": ["src"],
+  "exclude": ["dist", "build", "node_modules"]
+}
