+++ packages/components/entry/src/index.ts
@@ -8,3 +8,4 @@ export * from "@brifui/form";
 export * from "@brifui/text";
 export * from "@brifui/table";
 export * from "@brifui/toast";
+export * from "@brifui/avatar";
